Refactor route agent cni module to pkg/cni
...as it's used by components outside of the route agent
