+++ pkg/routeagent_driver/handlers/kubeproxy/kp_packetfilter.go
@@ -29,7 +29,6 @@ import (
 	"github.com/submariner-io/submariner/pkg/event"
 	"github.com/submariner-io/submariner/pkg/netlink"
 	"github.com/submariner-io/submariner/pkg/packetfilter"
-	cniapi "github.com/submariner-io/submariner/pkg/routeagent_driver/cni"
 	"github.com/submariner-io/submariner/pkg/vxlan"
 	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
 	"k8s.io/utils/set"
@@ -52,7 +51,7 @@ type SyncHandler struct {
 	vxlanDevice            *vxlan.Interface
 	vxlanGwIP              *net.IP
 	hostname               string
-	cniIface               *cniapi.Interface
+	cniIface               *cni.Interface
 	defaultHostIface       *net.Interface
 	activeEndpointHostname string
 }
@@ -106,7 +105,7 @@ func (kp *SyncHandler) Init() error {
 		return errors.Wrapf(err, "Unable to find the default interface on host: %s", kp.hostname)
 	}
 
-	cniIface, err := cniapi.Discover(kp.localClusterCidr[0])
+	cniIface, err := cni.Discover(kp.localClusterCidr[0])
 	if err == nil {
 		// Configure CNI Specific changes
 		kp.cniIface = cniIface
